requirements.txt
torch==1.8.0+cu111
torchvision==0.9.0+cu111
torchaudio==0.8.0
mmcv-full==1.5.3
mmdet==2.27.0
mmpose==0.28.1
mmcls==0.25.0
smplx==0.1.28
pytorch3d==0.7.2
opencv-python==4.7.0.68
numpy==1.23.5
scipy==1.10.0
matplotlib==3.6.2
tqdm==4.64.1
pyyaml==6.0
timm==0.6.12
einops==0.6.0
tensorboard==2.11.2
tensorboardx==2.6
pycocotools==2.0.6
h5py==3.7.0
trimesh==3.18.0
pyrender==0.1.45
vedo==2022.4.2
vtk==9.2.5
yacs==0.1.8
addict==2.4.0 